(PC-31640) feat(location): force reset initial venues in store when opening venue map location modal to avoid bad behavior when switching from search results to home when you decide not share location and press button to map access
