+++ src/features/home/components/modules/TrendsModule.tsx
@@ -6,7 +6,6 @@ import { Trend } from 'features/home/components/Trend'
 import { TrendBlock, TrendNavigationProps } from 'features/home/types'
 import { VenueMapLocationModal } from 'features/location/components/VenueMapLocationModal'
 import { useShouldDisplayVenueMap } from 'features/venueMap/hook/useShouldDisplayVenueMap'
-import { useInitialVenuesActions } from 'features/venueMap/store/initialVenuesStore'
 import { useSelectedVenueActions } from 'features/venueMap/store/selectedVenueStore'
 import { analytics } from 'libs/analytics'
 import { ContentTypes } from 'libs/contentful/types'
@@ -40,7 +39,6 @@ export const TrendsModule = ({ index, moduleId, homeEntryId, items }: Trends) =>
     hideModal: hideVenueMapLocationModal,
   } = useModal()
   const { removeSelectedVenue } = useSelectedVenueActions()
-  const { setInitialVenues } = useInitialVenuesActions()
 
   const isSmallScreen = width < 375
   const shouldOpenMapDirectly = selectedLocationMode !== LocationMode.EVERYWHERE && !isWeb
@@ -72,7 +70,6 @@ export const TrendsModule = ({ index, moduleId, homeEntryId, items }: Trends) =>
         enableNavigate: shouldOpenMapDirectly,
         onBeforeNavigate: () => {
           removeSelectedVenue()
-          setInitialVenues([])
           handleLogTrendsBlockClicked(props)
           analytics.logConsultVenueMap({ from: 'trend_block' })
           if (!shouldOpenMapDirectly) showVenueMapLocationModal()
